Coffee_machine_mini_project
This Python code uses modules like "art" and "game_data" to display ASCII art and provide data for the game. It imports the "random" module for random selection. The "replit" module is used for clearing the console. The code uses variables, conditional statements, loops, and user input to create a guessing game where players compare follower counts of celebrities.